--- workers/unity/Packages/io.improbable.gdk.buildsystem/Util/CommandlineParser.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aa29fe14f7ce4d2687ed1286578cda8e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
